Added method to resolve profile artifact on ProfileRepository. Cleanup of GrailsDependencyVersions
